+++ quickshell/Bar.qml
@@ -253,8 +253,6 @@ PanelWindow {
       radius: root.appConfig.cfg.barRadius
       padding: root.appConfig.cfg.pillPadding
       contentSpacing: root.appConfig.cfg.contentSpacing
-      glowSize: root.appConfig.cfg.barGlowSize
-      glowOpacity: root.appConfig.cfg.barGlowOpacity
       lighten: root.appConfig.cfg.barLighten
       bgOpacity: root.appConfig.cfg.barBgOpacity
       borderWidth: root.appConfig.cfg.barBorderWidth
@@ -274,8 +272,6 @@ PanelWindow {
       radius: root.appConfig.cfg.barRadius
       padding: root.appConfig.cfg.pillPadding
       contentSpacing: root.appConfig.cfg.contentSpacing
-      glowSize: root.appConfig.cfg.barGlowSize
-      glowOpacity: root.appConfig.cfg.barGlowOpacity
       lighten: root.appConfig.cfg.barLighten
       bgOpacity: root.appConfig.cfg.barBgOpacity
       borderWidth: root.appConfig.cfg.barBorderWidth
@@ -299,8 +295,6 @@ PanelWindow {
       radius: root.appConfig.cfg.barRadius
       padding: root.appConfig.cfg.pillPadding
       contentSpacing: root.appConfig.cfg.contentSpacing
-      glowSize: root.appConfig.cfg.barGlowSize
-      glowOpacity: root.appConfig.cfg.barGlowOpacity
       lighten: root.appConfig.cfg.barLighten
       bgOpacity: root.appConfig.cfg.barBgOpacity
       borderWidth: root.appConfig.cfg.barBorderWidth